requirements.txt
# onnxruntime==1.17.1 # chromadb dependency: on Mac use `conda install onnxruntime -c conda-forge`
# For Windows users, install Microsoft Visual C++ Build Tools first
# install onnxruntime before installing `chromadb`
# Backend-specific dependencies
Flask==2.3.2  # For the web app
python-dotenv==1.0.1 # For reading environment variables stored in .env file
langchain==0.2.2
langchain-community==0.2.3
langchain-openai==0.1.8
unstructured==0.14.4
chromadb==0.5.0
openai==1.31.1
tiktoken==0.7.0
sentence_transformers==3.0.1
flask_socketio